Cleanup nc2.util, pass 2. Move DiskCache, DiskCacje2, EscapeStrings to internal. Move URLnaming to util.net. Remove deprecated Misc methods.
