+++ cdm/core/src/main/java/thredds/client/catalog/Catalog.java
@@ -10,7 +10,7 @@ import thredds.client.catalog.builder.CatalogBuilder;
 import thredds.client.catalog.builder.CatalogRefBuilder;
 import thredds.client.catalog.builder.DatasetBuilder;
 import ucar.nc2.time.CalendarDate;
-import ucar.nc2.util.URLnaming;
+import ucar.nc2.util.net.URLnaming;
 import javax.annotation.concurrent.Immutable;
 import java.net.URI;
 import java.net.URISyntaxException;
