+++ cdm/core/src/main/java/ucar/nc2/NetcdfFiles.java
@@ -31,8 +31,8 @@ import ucar.nc2.internal.util.StringLocker;
 import ucar.nc2.iosp.AbstractIOServiceProvider;
 import ucar.nc2.iosp.IOServiceProvider;
 import ucar.nc2.util.CancelTask;
-import ucar.nc2.util.DiskCache;
-import ucar.nc2.util.EscapeStrings;
+import ucar.nc2.internal.util.DiskCache;
+import ucar.nc2.internal.util.EscapeStrings;
 import ucar.nc2.util.IO;
 import ucar.unidata.io.UncompressInputStream;
 import ucar.unidata.io.bzip2.CBZip2InputStream;
