+++ cdm/core/src/main/java/ucar/nc2/internal/util/DiskCache2.java
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) 1998-2018 John Caron and University Corporation for Atmospheric Research/Unidata
+ * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
-package ucar.nc2.util;
+package ucar.nc2.internal.util;
 
 import ucar.nc2.time.CalendarDate;
 import ucar.unidata.util.StringUtil2;
@@ -27,8 +27,6 @@ import java.util.*;
  * <li>through the system property "user.dir" if it exists
  * <li>to the current working directory
  * </ol>
- * 
- * TODO will move in ver 6
  */
 public class DiskCache2 {
   private static final org.slf4j.Logger cacheLog = org.slf4j.LoggerFactory.getLogger("cacheLogger");
