+++ cdm/core/src/main/java/ucar/nc2/internal/util/EscapeStrings.java
@@ -1,6 +1,9 @@
-/* Copyright Unidata */
+/*
+ * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
+ * See LICENSE for license information.
+ */
 
-package ucar.nc2.util;
+package ucar.nc2.internal.util;
 
 import com.google.re2j.Matcher;
 import com.google.re2j.Pattern;
