+++ cdm/core/src/main/java/ucar/nc2/util/net/URLnaming.java
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) 1998-2018 John Caron and University Corporation for Atmospheric Research/Unidata
+ * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
-package ucar.nc2.util;
+package ucar.nc2.util.net;
 
 import ucar.unidata.util.StringUtil2;
 import java.io.File;
