+++ cdm/core/src/test/java/ucar/nc2/util/DiskCache2Test.java
@@ -9,6 +9,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import ucar.nc2.internal.util.DiskCache2;
 
 /**
  * @author cwardgar
